WIP move stack and queue file into folder with README
