+++ python/code_challenges/stack_queue_animal_shelter.py
@@ -1,6 +1,3 @@
-from data_structures.queue import Queue
-
-
 class AnimalShelter:
     pass
