+++ python/code_challenges/stack_queue_brackets.py
@@ -1,5 +1,2 @@
-from data_structures.queue import Queue
-
-
 def multi_bracket_validation():
     pass
