+++ python/code_challenges/stack_queue_pseudo.py
@@ -1,5 +1,2 @@
-from data_structures.stack import Stack
-
-
 class PseudoQueue:
     pass
