+++ python/data_structures/kary_tree.py
@@ -1,4 +1,4 @@
-from data_structures.queue import Queue
+from data_structures.stack_and_queue.queue import Queue
 
 
 class KaryTree:
