+++ python/data_structures/stack_and_queue/README.md
@@ -48,11 +48,46 @@ Stack:
 - `push`: Adds a new node with that value to the top of the stack
 - `pop`: Removes the node from the top of the stack and returns its value. Raises exception when called on an empty stack
 - `peek`: Returns the value of the node located at the top of the stack. Raises exception when called on an empty stack
-- `is_empty`: Returns boolean indicating whether or not the stack is empty
+- `is_empty`: Returns boolean indicating whether the stack is empty
 
 Queue:
 
 - `enqueue`: Adds a new node with that value to the back of the queue with an O(1) Time performance
 - `dequeue`: Removes the node from the front of the queue and returns the value from node from the front of the queue. Throws exception when called on an empty queue
-- `peek`: Returns the value of the node located at the front of the queue. Raises exception when called on an empty stack
-- `is_empty`: Returns boolean indicating whether or not the queue is empty
+- `peek`: Returns the value of the node located at the front of the queue. Raises exception when called on an empty
+  queue
+- `is_empty`: Returns boolean indicating whether the queue is empty
+
+## Setup
+
+To run the tests for this code challenge, make sure you `cd` into the `python` directory first.
+
+Then create a virtual environment:
+
+```bash
+python3 -m venv .venv
+```
+
+Then activate the virtual environment:
+
+```bash
+source .venv/bin/activate
+```
+
+Then install the dependencies:
+
+```bash
+pip install -r requirements.txt
+```
+
+To run the tests for this code challenge:
+
+```bash
+pytest -v tests/data_structures/test_stack.py tests/data_structures/test_queue.py
+```
+
+To deactivate the virtual environment:
+
+```bash
+deactivate
+```
