+++ python/data_structures/stack_and_queue/stack.py
