+++ python/tests/code_challenges/test_tree_intersection.py
@@ -1,7 +1,7 @@
 import pytest
 from code_challenges.tree_intersection import tree_intersection
 from data_structures.binary_tree import BinaryTree, Node
-from data_structures.queue import Queue
+from data_structures.stack_and_queue.queue import Queue
 
 
 def test_exists():
