+++ python/tests/data_structures/test_stack.py
@@ -1,5 +1,5 @@
 import pytest
-from data_structures.stack import Stack
+from data_structures.stack_and_queue.stack import Stack
 from data_structures.invalid_operation_error import InvalidOperationError
